requirements.txt
# Core Flask
Flask==3.0.0
Flask-Cors==4.0.0
Flask-SQLAlchemy==3.1.1

# Database
PyMySQL==1.1.0

# Image processing
pillow==10.2.0
numpy>=1.26.0,<2.1.0

# PDF generation
reportlab==4.0.8
qrcode[pil]==7.4.2

# Security
Werkzeug==3.0.1

# Utilities
python-dotenv==1.0.0
google-generativeai==0.8.6

# ML
tensorflow==2.18.1
keras>=3.5.0
